# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application Programming Interface (API) for Pasiap Apps",
        "description": "L5 Swagger OpenAPI documentation for Pasiap Apps",
        "contact": {
            "email": "burhanburdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/pasiap/public/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "description": "Get all reports",
                "operationId": "getReports",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reports"
                ],
                "description": "Create new report",
                "operationId": "createReport",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "cat_id": {
                                        "type": "string"
                                    },
                                    "reported_by": {
                                        "type": "string"
                                    },
                                    "lat": {
                                        "type": "string"
                                    },
                                    "long": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "integer"
                                    },
                                    "photo": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "description": "Get detail report by ID",
                "operationId": "getReportById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the data to show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reports"
                ],
                "description": "Update report by ID",
                "operationId": "updateReport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the data to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "cat_id": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/status/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "description": "Get detail report by ID status",
                "operationId": "getReportByStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID status of the data to show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reports"
                ],
                "description": "Update report status by Report ID",
                "operationId": "updateReportStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the data to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "taken_by": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/handler/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "description": "Get detail report by ID handler reports",
                "operationId": "getReportByHandler",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID handler reports of the data to show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/request/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "description": "Get detail report by ID requestor reports",
                "operationId": "getReportByRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID requestor reports of the data to show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}